# Reroute algebraic functions for `Symmetric`/`Hermitian` through triangular