@araujoms 2024-01-17
This PR exceeds my most optimistic expectations. Note that it doesn't work for mismatched uplo, though, because that will be wrapped by `Adjoint`, and that falls back to the non-symmetric `+`. Example:

```julia
function matrix_sum(T)
   d = 2
   a = Hermitian(Matrix{T}(undef,d,d),:U)
   b = Hermitian(Matrix{T}(undef,d,d),:L)
   for i=1:d
       for j=i:d
           a.data[i,j] = T(1)
           b.data[j,i] = T(1)
       end
   end
   a+b
end    
```