@jishnub 2024-01-18
Yes, fixing the mismatched `uplo` case would require https://github.com/JuliaLang/julia/pull/52530