@jishnub 2024-01-29
The addition/subtraction with mismatched `uplo` is fixed now:
```julia
julia> matrix_sum(BigFloat)
2×2 Hermitian{BigFloat, Matrix{BigFloat}}:
 2.0  2.0
 2.0  2.0
```